+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed example for macos.
 * Updated Flutter to 3.3.7
 * Updated libMediaInfo from v21.09 to v22.09
+* Changed exception name from MediaInfoInstanceHasExist to MediaInfoInstanceHasExists
 
 ## 0.0.1+1